+++ CHANGES.md
@@ -10,6 +10,8 @@
 
 ### Changed
 
+  - Improve diff representation of modified record types (#109, @azzsal)
+
 ### Deprecated
 
 ### Fixed